refactor class Exception
replace name of Exception to TransaccionCompletaException
