+++ lib/onepay/exceptions/RefundCreateException.php
@@ -1,11 +1,11 @@
 <?php
 namespace Transbank\Onepay\Exceptions;
 /**
- * 
+ *
  * class RefundCreateException
  * Model object for raising exceptions when a Refund::create transaction fails.
- * 
- * @package Transbank\Onepay\Exceptions
+ *
+ * @package Transbank\Onepay\TransaccionCompletaException
  */
 class RefundCreateException extends TransbankException {
