+++ lib/transaccion_completa/exceptions/InscriptionFinishException.php
@@ -4,17 +4,17 @@
  * Class InscriptionFinishException
  *
  * @category
- * @package Transbank\TransaccionCompleta\Exceptions
+ * @package Transbank\TransaccionCompleta\TransaccionCompletaException
  *
  */
 
 
 namespace Transbank\TransaccionCompleta\Exceptions;
 
-use Transbank\TransaccionCompleta\Exceptions;
+use Transbank\TransaccionCompleta\Exceptions\TransaccionCompletaException;
 
 
-class InscriptionFinishException extends Exceptions
+class InscriptionFinishException extends TransaccionCompletaException
 {
     public function __construct($message = self::DEFAULT_MESSAGE, $code = 0, $previous = null)
     {
