+++ lib/transaccion_completa/exceptions/InscriptionStartException.php
@@ -1,9 +1,9 @@
 <?php
 namespace Transbank\TransaccionCompleta\Exceptions;
 
-use Transbank\TransaccionCompleta\Exceptions;
+use Transbank\TransaccionCompleta\Exceptions\TransaccionCompletaException;
 
-class InscriptionStartException extends Exceptions
+class InscriptionStartException extends TransaccionCompletaException
 {
     public function __construct($message = self::DEFAULT_MESSAGE, $code = 0, $previous = null)
     {
